LoadingBar magics sprinkled over encrypt / decrypt folder
The socket.js file now populates the electron.event variable (only at /index at the moment). This allows virtually all scripts to use the electron variable and gain access to electron.event which in turn allows to easily use the ipcRenderer for sending instant messages from the renderer to the main process.

This nifty feature allowed me to implement a loadingbar for encrypting and decrypting files. It's still WIP.